Initial React v0.14 definitions
