+++ react/react-addons-update.d.ts
@@ -1,0 +1,35 @@
+// Type definitions for React v0.14 (react-addons-update)
+// Project: http://facebook.github.io/react/
+// Definitions by: Asana <https://asana.com>, AssureSign <http://www.assuresign.com>, Microsoft <https://microsoft.com>
+// Definitions: https://github.com/borisyankov/DefinitelyTyped
+
+/// <reference path="react.d.ts" />
+
+declare namespace __React {
+    namespace __Addons {
+        interface UpdateSpecCommand {
+            $set?: any;
+            $merge?: {};
+            $apply?(value: any): any;
+        }
+    
+        interface UpdateSpecPath {
+            [key: string]: UpdateSpec;
+        }
+    
+        type UpdateSpec = UpdateSpecCommand | UpdateSpecPath;
+    
+        interface UpdateArraySpec extends UpdateSpecCommand {
+            $push?: any[];
+            $unshift?: any[];
+            $splice?: any[][];
+        }
+        
+        export function update(value: any[], spec: UpdateArraySpec): any[];
+        export function update(value: {}, spec: UpdateSpec): any;
+    }
+}
+
+declare module "react-addons-update" {
+    export = __React.__Addons.update;
+}
